refactor: separate personalization query hook and optimize display name update path
Agent-Logs-Url: https://github.com/palewire/LibreChat/sessions/5e586230-65fe-4dc1-a0af-f6365b1bc840
