+++ client/src/data-provider/Personalization/index.ts
@@ -1,0 +1,1 @@
+export * from './queries';
